screenrc: add quick-n-dirty hack to generate stty call for use with serial terminals
Due to the nature of using screen with serial terminals the exec
command will actually execute locally, and get the size of the
terminal using the resize command that ships with xterm.

Unfortunately it doesn't insert the generated values, it only displays
them, requiring you to copy and paste.

TODO: * Needs to actually insert the text
      * Get the sizes of the terminal using something else than
      resize. Either via using escape codes directly, or perhaps
      'tput lines' and 'tput cols'.