screenrc update
Let's try without the running of .bash_profile on startup and see if
things are still ok? I've been running this way on usbee-ubuntu for a
long time at the very least. Don't really remember why I first
uncommented it...

Some minor cleanup

Add a commented potentially useful slowpaste setting just to remember
myself that this is an alternative when working with u-boot (when
needed, just uncomment the line or run the command interactively in a
screen session)