waybar: config updated to work better with newer versions
This is based on a newer version of the template config.
Also I added some workarounds where I had font issues on
usbee-ubuntu for some of the icons (maybe I should just have a
usbee-ubuntu-specific config for this reason? I haven't had as many
issues with this on other machines)